# src/docs/openApi.yaml
openapi: 3.0.3
info:
  title: Order Management API
  version: 1.0.0
  description: Swagger specification for the Order CRUD operations built around the Order model.
servers:
  - url: http://localhost:3000
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: List all users
      operationId: listUsers
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags:
        - Users
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
            examples:
              user:
                summary: User example
                value:
                  name: Jane Doe
                  email: jane.doe@example.com
                  password: correct-horse-battery-staple
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid user payload
        '409':
          description: A user with this email already exists
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get a user by id
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: The requested user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    put:
      tags:
        - Users
      summary: Update an existing user
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid user payload
        '404':
          description: User not found
        '409':
          description: A user with this email already exists
    delete:
      tags:
        - Users
      summary: Delete a user by id
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
...
components:
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      description: The user identifier
      schema:
        type: string
  schemas:
    User:
      type: object
      required:
        - id
        - name
        - email
      properties:
        id:
          type: string
          example: USR-1001
        name:
          type: string
          example: Jane Doe
        email:
          type: string
          format: email
          example: jane.doe@example.com
    UserCreateRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          example: Jane Doe
        email:
          type: string
          format: email
          example: jane.doe@example.com
        password:
          type: string
          format: password
          example: correct-horse-battery-staple
    UserUpdateRequest:
      type: object
      required:
        - id
        - name
        - email
        - password
      properties:
        id:
          type: string
          example: USR-1001
        name:
          type: string
          example: Jane Doe
        email:
          type: string
          format: email
          example: jane.doe@example.com
        password:
          type: string
          format: password
          example: correct-horse-battery-staple
...